# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Robot Shop",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8000",
    "paths": {
        "/admin/apis": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "搜索时支持分组名、名称、路径和方法筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "获取接口列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分组名",
                        "name": "group",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "路径",
                        "name": "path",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "方法",
                        "name": "method",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.ApiSearchResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建一个新的接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "创建接口",
                "operationId": "ApiCreate",
                "parameters": [
                    {
                        "description": "接口数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.ApiRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/admin/apis/{id}": {
            "get": {
                "description": "获取指定ID的接口信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "获取接口",
                "operationId": "GetApi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接口ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.ApiResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新接口数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "更新接口",
                "operationId": "ApiUpdate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接口ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "接口数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.ApiRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定ID的接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API"
                ],
                "summary": "删除接口",
                "operationId": "ApiDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "接口ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/admin/menus": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取所有菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "获取菜单列表",
                "operationId": "ListMenus",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.MenuSearchResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建一个新的菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "创建菜单",
                "operationId": "MenuCreate",
                "parameters": [
                    {
                        "description": "菜单数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.MenuRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
        "/admin/menus/{id}": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "更新菜单数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "更新菜单",
                "operationId": "MenuUpdate",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "菜单数据",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.MenuRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "删除指定ID的菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "删除菜单",
                "operationId": "MenuDelete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "菜单ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/backend_api_v1.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Api": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "创建时间",
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                },
                "group": {
                    "description": "分组名",
                    "type": "string",
                    "example": "PermManagement"
                },
                "id": {
                    "description": "ID",
                    "type": "integer",
                    "example": 1
                },
                "method": {
                    "description": "方法",
                    "type": "string",
                    "example": "GET"
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "example": "MenuList"
                },
                "path": {
                    "description": "路径",
                    "type": "string",
                    "example": "/v1/xxx"
                },
                "updatedAt": {
                    "description": "更新时间",
                    "type": "string",
                    "example": "2006-01-02 15:04:05"
                }
            }
        },
        "ApiList": {
            "type": "object",
            "properties": {
                "groups": {
                    "description": "分组名列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "list": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Api"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "Menu": {
            "type": "object",
            "properties": {
                "component": {
                    "description": "绑定的组件",
                    "type": "string"
                },
                "hideInMenu": {
                    "description": "是否保活",
                    "type": "boolean"
                },
                "icon": {
                    "description": "图标，使用字符串表示",
                    "type": "string"
                },
                "id": {
                    "description": "唯一id，使用整数表示",
                    "type": "integer"
                },
                "keepAlive": {
                    "description": "是否保活",
                    "type": "boolean"
                },
                "locale": {
                    "description": "本地化标识",
                    "type": "string"
                },
                "name": {
                    "description": "同路由中的name，唯一标识",
                    "type": "string"
                },
                "parentId": {
                    "description": "父级菜单的id，使用整数表示",
                    "type": "integer"
                },
                "path": {
                    "description": "地址",
                    "type": "string"
                },
                "redirect": {
                    "description": "重定向地址",
                    "type": "string"
                },
                "title": {
                    "description": "展示名称",
                    "type": "string"
                },
                "updatedAt": {
                    "description": "是否保活",
                    "type": "string"
                },
                "url": {
                    "description": "iframe模式下的跳转url，不能与path重复",
                    "type": "string"
                },
                "weight": {
                    "description": "排序权重",
                    "type": "integer"
                }
            }
        },
        "MenuList": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Menu"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
...
        "backend_api_v1.ApiRequest": {
            "type": "object",
            "properties": {
                "group": {
                    "description": "分组名",
                    "type": "string",
                    "example": "PermManagement"
                },
                "method": {
                    "description": "方法",
                    "type": "string",
                    "example": "GET"
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "example": "MenuList"
                },
                "path": {
                    "description": "路径",
                    "type": "string",
                    "example": "/v1/xxx"
                }
            }
        },
        "backend_api_v1.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/Api"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "backend_api_v1.ApiSearchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/ApiList"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_api_v1.MenuRequest": {
            "type": "object",
            "properties": {
                "component": {
                    "description": "绑定的组件",
                    "type": "string"
                },
                "hideInMenu": {
                    "description": "是否保活",
                    "type": "boolean"
                },
                "icon": {
                    "description": "图标，使用字符串表示",
                    "type": "string"
                },
                "keepAlive": {
                    "description": "是否保活",
                    "type": "boolean"
                },
                "locale": {
                    "description": "本地化标识",
                    "type": "string"
                },
                "name": {
                    "description": "同路由中的name，唯一标识",
                    "type": "string"
                },
                "parentId": {
                    "description": "父级菜单的id，使用整数表示",
                    "type": "integer"
                },
                "path": {
                    "description": "地址",
                    "type": "string"
                },
                "redirect": {
                    "description": "重定向地址",
                    "type": "string"
                },
                "title": {
                    "description": "展示名称",
                    "type": "string"
                },
                "url": {
                    "description": "iframe模式下的跳转url，不能与path重复",
                    "type": "string"
                },
                "weight": {
                    "description": "排序权重",
                    "type": "integer"
                }
            }
        },
        "backend_api_v1.MenuSearchResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/MenuList"
                },
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "backend_api_v1.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errorCode": {
                    "description": "错误码",
                    "type": "integer"
                },
                "errorMessage": {
                    "description": "报错信息",
                    "type": "string"
                },
                "errorShowType": {
                    "description": "前端展示方式",
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}